separated sections into their own classes
